Relocate internal packages and update imports
- Moved `pkg/internal/` to `internal/` to align with
  [Go's internal package conventions](https://go.dev/doc/modules/layout#package-or-command-with-supporting-packages).
- Updated all relevant import paths to reflect the new package structure.
- Adjusted `make/test-unit.mk` to include the new directory structure in test coverage.
- Ensured all tests and mock data are updated to work with the new structure.
