+++ internal/cyberark/dataupload/dataupload.go
