+++ internal/cyberark/dataupload/mock.go
