+++ internal/cyberark/identity/advance_authentication_test.go
@@ -7,7 +7,7 @@ import (
 	"k8s.io/klog/v2"
 	"k8s.io/klog/v2/ktesting"
 
-	"github.com/jetstack/preflight/pkg/internal/cyberark/servicediscovery"
+	"github.com/jetstack/preflight/internal/cyberark/servicediscovery"
 
 	_ "k8s.io/klog/v2/ktesting/init"
 )
