+++ internal/cyberark/identity/authenticated_http_client.go
