+++ internal/cyberark/identity/mock.go
