+++ internal/cyberark/identity/testdata/advance_authentication_success.json
