+++ internal/cyberark/identity/testdata/start_authentication_bad_user_session_id.json
