+++ internal/cyberark/identity/testdata/start_authentication_success.json
