+++ internal/cyberark/identity/testdata/start_authentication_success_multiple_mechanisms.json
