+++ internal/cyberark/identity/testdata/start_authentication_success_no_up_mechanism.json
