+++ internal/cyberark/servicediscovery/discovery.go
