+++ internal/cyberark/servicediscovery/discovery_test.go
