+++ internal/cyberark/servicediscovery/mock.go
