+++ internal/cyberark/servicediscovery/testdata/discovery_success.json.template
