+++ pkg/client/client_cyberark.go
@@ -9,8 +9,8 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	"github.com/jetstack/preflight/api"
-	"github.com/jetstack/preflight/pkg/internal/cyberark"
-	"github.com/jetstack/preflight/pkg/internal/cyberark/dataupload"
+	"github.com/jetstack/preflight/internal/cyberark"
+	"github.com/jetstack/preflight/internal/cyberark/dataupload"
 	"github.com/jetstack/preflight/pkg/version"
 )
